requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
openai>=1.0.0
databricks-sdk>=0.17.0
mlflow>=3.0.0
